fix(skills): cap skill-library fetch concurrency to avoid raw CDN 429
listSkillLibrary fetched every SKILL.md in parallel via Promise.all,
bursting past raw.githubusercontent.com's per-IP throttle and randomly
dropping descriptions (empty-string descriptions scattered across the
list). PR #66 fixed the preset side the same way but left the skill
library uncapped.

Extract the mapWithConcurrency helper PR #66 added privately in
preset-store.ts into the shared content-source/types.ts, and use it in
listSkillLibrary with a concurrency limit of 5 (matching presets). Both
sites now import the shared helper instead of duplicating it.
